split up some test code
